Sync safes with our db and ui
